Reconfigure the tabs into respective pages * course.description => TOC tab * course.keypoints, client testimonials => Promo * button to update the course
